refactor: extract UniversalSelector into its own module
freecad_mcp_handler.py was at 2039 lines against its documented
700-750 target — mixing genuine socket-server infrastructure with
standalone business logic that belongs elsewhere. UniversalSelector
(the interactive edge/face selection manager) has no dependency on
the socket server itself, only its own pending_operations dict plus
FreeCADGui.Selection/FreeCAD.ActiveDocument — first of a 3-part
extraction pass (2039 -> 1897 lines).

Two tests patched FreeCADGui on the wrong module post-move (the
class's methods close over universal_selector's own module-global,
not freecad_mcp_handler's) — fixed to patch the correct module.
